+++ packages/authhero/src/components/ForgotPasswordPage.tsx
@@ -1,0 +1,52 @@
+import type { FC } from "hono/jsx";
+import Layout from "./Layout";
+import { VendorSettings } from "authhero";
+import i18next from "i18next";
+import ErrorMessage from "./ErrorMessage";
+import DisabledSubmitButton from "./DisabledSubmitButton";
+import Form from "./Form";
+import { GoBack } from "./GoBack";
+
+type Props = {
+  error?: string;
+  vendorSettings: VendorSettings;
+  email?: string;
+  state: string;
+};
+
+const ForgotPasswordPage: FC<Props> = (parms) => {
+  const { error, vendorSettings, email, state } = parms;
+
+  return (
+    <Layout
+      title={i18next.t("forgot_password_title")}
+      vendorSettings={vendorSettings}
+    >
+      <div className="mb-4 text-lg font-medium sm:text-2xl">
+        {i18next.t("forgot_password_title")}
+      </div>
+      <div className="mb-6 text-gray-300">
+        {i18next.t("forgot_password_description")}
+      </div>
+      <div className="flex flex-1 flex-col justify-center">
+        <Form className="pt-2">
+          <input
+            type="email"
+            name="username"
+            placeholder={i18next.t("email_placeholder")}
+            className="mb-2 w-full rounded-lg bg-gray-100 px-4 py-5 text-base placeholder:text-gray-300 dark:bg-gray-600 md:text-base"
+            value={email}
+            disabled={!!email}
+          />
+          {error && <ErrorMessage>{error}</ErrorMessage>}
+          <DisabledSubmitButton className="sm:mt-4">
+            {i18next.t("forgot_password_cta")}
+          </DisabledSubmitButton>
+        </Form>
+        <GoBack state={state} />
+      </div>
+    </Layout>
+  );
+};
+
+export default ForgotPasswordPage;