+++ packages/kysely/src/hooks/list.ts
@@ -8,7 +8,11 @@ import getCountAsInt from "../utils/getCountAsInt";
 export function list(db: Kysely<Database>) {
   return async (
     tenant_id: string,
-    params: ListParams,
+    params: ListParams = {
+      page: 0,
+      per_page: 50,
+      include_totals: false,
+    },
   ): Promise<ListHooksResponse> => {
     let query = db.selectFrom("hooks").where("hooks.tenant_id", "=", tenant_id);